requirements.txt
aiohttp
aiosqlite
discord.py
google-api-python-client
google-auth-httplib2
google-auth-oauthlib